+++ @xen-orchestra/lite/CHANGELOG.md
@@ -1,5 +1,9 @@
 # ChangeLog
 
+## **next**
+
+- [Host] Add dashboard view (PR [#8504](https://github.com/vatesfr/xen-orchestra/pull/8504))
+
 ## **0.9.1** (2025-04-02)
 
 - Fix blank page in production mode (PR [#8471](https://github.com/vatesfr/xen-orchestra/pull/8471))