+++ @xen-orchestra/lite/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - [Host/Header] Add master host icon on host header (PR [#8512](https://github.com/vatesfr/xen-orchestra/pull/8512))
 - [Host] Add dashboard view (PR [#8504](https://github.com/vatesfr/xen-orchestra/pull/8504))
+- [Host] Add system view (PR [#8536](https://github.com/vatesfr/xen-orchestra/pull/8536))
 
 ## **0.9.1** (2025-04-02)